Adapted current unit tests for recent changes.
