+++ tests/test_elder_spawns.py
@@ -1,0 +1,11 @@
+import brownie
+
+
+def test_get_elder_spawns_invalid_id(accounts, crypto_champions):
+    with brownie.reverts("dev: Given id is not valid."):
+        crypto_champions.getElderSpawnsAmount(crypto_champions.currentRound(), 0)
+
+
+def test_get_elder_spawns_wrong_round(accounts, crypto_champions, mint_first_elder):
+    with brownie.reverts("dev: Invalid round."):
+        crypto_champions.getElderSpawnsAmount(crypto_champions.currentRound() + 1, 1)
