+++ tests/test_pricing_mechanics.py
@@ -19,37 +19,37 @@ def test_get_hero_mint_price_invalid_id(accounts, crypto_champions):
 
 
 def test_mint_elder_rewards_pool(accounts, crypto_champions, mint_first_elder):
-    assert crypto_champions.rewardsPoolAmount() == crypto_champions.elderMintPrice()
+    assert crypto_champions.devFund() == crypto_champions.elderMintPrice()
 
 
 def test_mint_elder_rewards_pool_refund(accounts, crypto_champions, create_eth_affinity):
     crypto_champions.mintElderSpirit(0, 0, "ETH", {"from": accounts[0], "value": crypto_champions.elderMintPrice() + 1000})
-    assert crypto_champions.rewardsPoolAmount() == crypto_champions.elderMintPrice()
+    assert crypto_champions.devFund() == crypto_champions.elderMintPrice()
     
 
 def test_mint_hero_rewards_pool(accounts, crypto_champions, mint_first_elder, set_phase_to_mint_hero):
     mintAmount = crypto_champions.getHeroMintPrice(crypto_champions.currentRound(), 1)
-    rewardsPoolAmount = crypto_champions.elderMintPrice() + mintAmount * 0.75 # hardcoded for now
+    rewardsPoolAmount = mintAmount * 0.5 # hardcoded for now
     crypto_champions.mintHero(1, "Name", {"from": accounts[0], "value": mintAmount})
     assert crypto_champions.rewardsPoolAmount() == rewardsPoolAmount
 
 
 def test_mint_hero_rewards_pool_refund(accounts, crypto_champions, mint_first_elder, set_phase_to_mint_hero):
     mintAmount = crypto_champions.getHeroMintPrice(crypto_champions.currentRound(), 1)
-    rewardsPoolAmount = crypto_champions.elderMintPrice() + mintAmount * 0.75 # hardcoded for now
+    rewardsPoolAmount = mintAmount * 0.5 # hardcoded for now
     crypto_champions.mintHero(1, "Name", {"from": accounts[0], "value": mintAmount + 1000})
     assert crypto_champions.rewardsPoolAmount() == rewardsPoolAmount
 
 
 def test_mint_hero_dev_pool(accounts, crypto_champions, mint_first_elder, set_phase_to_mint_hero):
     mintAmount = crypto_champions.getHeroMintPrice(crypto_champions.currentRound(), 1)
-    devFundAmount = mintAmount * 0.05 # hardcoded for now
+    devFundAmount = crypto_champions.elderMintPrice() + mintAmount * 0.25 # hardcoded for now
     crypto_champions.mintHero(1, "Name", {"from": accounts[0], "value": mintAmount})
     assert crypto_champions.devFund() == devFundAmount
 
 
 def test_mint_hero_dev_pool_refund(accounts, crypto_champions, mint_first_elder, set_phase_to_mint_hero):
     mintAmount = crypto_champions.getHeroMintPrice(crypto_champions.currentRound(), 1)
-    devFundAmount = mintAmount * 0.05 # hardcoded for now
+    devFundAmount = crypto_champions.elderMintPrice() + mintAmount * 0.25 # hardcoded for now
     crypto_champions.mintHero(1, "Name", {"from": accounts[0], "value": mintAmount + 1000})
     assert crypto_champions.devFund() == devFundAmount
